Move AddressBook to person package
