+++ src/main/java/seedu/vms/MainApp.java
@@ -15,12 +15,12 @@ import seedu.vms.commons.util.ConfigUtil;
 import seedu.vms.commons.util.StringUtil;
 import seedu.vms.logic.Logic;
 import seedu.vms.logic.LogicManager;
-import seedu.vms.model.AddressBook;
 import seedu.vms.model.Model;
 import seedu.vms.model.ModelManager;
-import seedu.vms.model.ReadOnlyAddressBook;
 import seedu.vms.model.ReadOnlyUserPrefs;
 import seedu.vms.model.UserPrefs;
+import seedu.vms.model.person.AddressBook;
+import seedu.vms.model.person.ReadOnlyAddressBook;
 import seedu.vms.model.util.SampleDataUtil;
 import seedu.vms.storage.AddressBookStorage;
 import seedu.vms.storage.JsonAddressBookStorage;
