+++ src/main/java/seedu/vms/storage/StorageManager.java
@@ -7,9 +7,9 @@ import java.util.logging.Logger;
 
 import seedu.vms.commons.core.LogsCenter;
 import seedu.vms.commons.exceptions.DataConversionException;
-import seedu.vms.model.ReadOnlyAddressBook;
 import seedu.vms.model.ReadOnlyUserPrefs;
 import seedu.vms.model.UserPrefs;
+import seedu.vms.model.person.ReadOnlyAddressBook;
 
 /**
  * Manages storage of AddressBook data in local storage.
