+++ src/test/java/seedu/vms/logic/LogicManagerTest.java
@@ -25,9 +25,9 @@ import seedu.vms.logic.parser.exceptions.ParseException;
 import seedu.vms.logic.parser.patient.PatientParser;
 import seedu.vms.model.Model;
 import seedu.vms.model.ModelManager;
-import seedu.vms.model.ReadOnlyAddressBook;
 import seedu.vms.model.UserPrefs;
 import seedu.vms.model.person.Person;
+import seedu.vms.model.person.ReadOnlyAddressBook;
 import seedu.vms.storage.JsonAddressBookStorage;
 import seedu.vms.storage.JsonUserPrefsStorage;
 import seedu.vms.storage.StorageManager;
