+++ src/test/java/seedu/vms/logic/commands/CommandTestUtil.java
@@ -15,9 +15,9 @@ import java.util.Map;
 import seedu.vms.commons.core.index.Index;
 import seedu.vms.logic.commands.exceptions.CommandException;
 import seedu.vms.logic.commands.patient.EditCommand;
-import seedu.vms.model.AddressBook;
 import seedu.vms.model.IdData;
 import seedu.vms.model.Model;
+import seedu.vms.model.person.AddressBook;
 import seedu.vms.model.person.NameContainsKeywordsPredicate;
 import seedu.vms.model.person.Person;
 import seedu.vms.testutil.EditPersonDescriptorBuilder;
