+++ src/test/java/seedu/vms/logic/commands/patient/EditCommandTest.java
@@ -19,10 +19,10 @@ import org.junit.jupiter.api.Test;
 import seedu.vms.commons.core.Messages;
 import seedu.vms.commons.core.index.Index;
 import seedu.vms.logic.commands.patient.EditCommand.EditPersonDescriptor;
-import seedu.vms.model.AddressBook;
 import seedu.vms.model.Model;
 import seedu.vms.model.ModelManager;
 import seedu.vms.model.UserPrefs;
+import seedu.vms.model.person.AddressBook;
 import seedu.vms.model.person.Person;
 import seedu.vms.testutil.EditPersonDescriptorBuilder;
 import seedu.vms.testutil.PersonBuilder;
