+++ src/test/java/seedu/vms/storage/JsonAddressBookStorageTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import seedu.vms.commons.exceptions.DataConversionException;
-import seedu.vms.model.AddressBook;
-import seedu.vms.model.ReadOnlyAddressBook;
+import seedu.vms.model.person.AddressBook;
+import seedu.vms.model.person.ReadOnlyAddressBook;
 
 public class JsonAddressBookStorageTest {
     private static final Path TEST_DATA_FOLDER = Paths.get("src", "test", "data", "JsonAddressBookStorageTest");
