+++ src/test/java/seedu/vms/storage/JsonSerializableAddressBookTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.vms.commons.exceptions.IllegalValueException;
 import seedu.vms.commons.util.JsonUtil;
-import seedu.vms.model.AddressBook;
+import seedu.vms.model.person.AddressBook;
 import seedu.vms.testutil.TypicalPersons;
 
 public class JsonSerializableAddressBookTest {
