+++ src/test/java/seedu/vms/storage/StorageManagerTest.java
@@ -11,9 +11,9 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import seedu.vms.commons.core.GuiSettings;
-import seedu.vms.model.AddressBook;
-import seedu.vms.model.ReadOnlyAddressBook;
 import seedu.vms.model.UserPrefs;
+import seedu.vms.model.person.AddressBook;
+import seedu.vms.model.person.ReadOnlyAddressBook;
 
 public class StorageManagerTest {
